Set calculator to be 9 digits and round numbers.
It rounds to 5 decimals and sets a limit of 9 digit to not overpopulate
the screen. If the number is too big it will show an alert and it will
truncate it. The results with decimals are rounded, if they still don't
fit in the screen, they are truncated to fit.